+++ web/usuaris/src/components/organisms/SideMenu.tsx
@@ -80,8 +80,7 @@ const CategoryStyled = styled.span`
 const ImgStyled = styled(Icon)`
   height: 30px;
   margin-right: ${dimensions.spacing.xxxs};
-
-  &[data-testid] {
+  &[data-testid] 
     content: attr(data-testid);
 `