+++ web/usuaris/src/components/organisms/UsersTable/UsersTable.tsx
@@ -3,7 +3,7 @@ import { ChangeEvent, FC, useEffect, useState } from 'react'
 import { useTranslation } from 'react-i18next'
 import { ColumnDef, createColumnHelper } from '@tanstack/react-table'
 import { Checkbox, Spinner, dimensions } from '@itacademy/ui'
-import { Table } from '../../molecules'
+import { DeleteConfirmationModal, Table } from '../../molecules'
 import { TFilters, TUserData, UserStatus } from '../../../types'
 import { icons } from '../../../assets/icons'
 import { useGetUsers, useUpdateUser } from '../../../hooks'
@@ -19,6 +19,7 @@ import {
   StatusStyled,
   TableContainer,
 } from './UsersTable.styles'
+import { UserRole } from '../../../types/types'
 
 type TUsersTable = {
   filtersSelected: TFilters | Record<string, never>
@@ -34,9 +35,14 @@ export const UsersTable: FC<TUsersTable> = ({ filtersSelected }) => {
   const [selectedStatus, setSelectedStatus] = useState<UserStatus | undefined>(
     undefined
   )
+
   const [selectedUsers, setSelectedUsers] = useState<TUserData[]>([])
 
   const { changeUserStatus } = useUpdateUser()
+
+  const [isModalOpen, setIsModalOpen] = useState<boolean>(false)
+
+  const [idToDelete, setIdToDelete] = useState<string>('')
 
   useEffect(() => {
     setFilters(filtersSelected)
@@ -96,7 +102,7 @@ export const UsersTable: FC<TUsersTable> = ({ filtersSelected }) => {
 
   const columHelper = createColumnHelper<TUserData>()
 
-  const columns: ColumnDef<TUserData, UserStatus>[] = [
+  const columns: ColumnDef<TUserData, string>[] = [
     columHelper.accessor('id', {
       header: '',
       cell: ({ row }) => {
@@ -217,6 +223,24 @@ export const UsersTable: FC<TUsersTable> = ({ filtersSelected }) => {
         )
       },
     }),
+
+    columHelper.accessor('role', {
+      header: `${t('Rol')}`,
+      cell: ({ row }) => {
+        const role: UserRole = row.getValue('role')
+        const status: UserStatus = row.getValue('status')
+        let isDisabled: boolean | undefined
+
+        if (selectedStatus && selectedStatus !== status) {
+          isDisabled = true
+        } else {
+          isDisabled = undefined
+        }
+
+        return <DisabledStyled disabled={isDisabled}>{t(role)}</DisabledStyled>
+      },
+    }),
+
     columHelper.display({
       id: 'actions',
       header: () => (
@@ -258,7 +282,16 @@ export const UsersTable: FC<TUsersTable> = ({ filtersSelected }) => {
             >
               {buttonTxt}
             </ButtonStyled>
-            <DeleteButton size="small" outline disabled={isDisabled}>
+            <DeleteButton
+              data-testid="delete-button"
+              size="small"
+              outline
+              disabled={isDisabled}
+              onClick={() => {
+                setIdToDelete(id)
+                setIsModalOpen(!isModalOpen)
+              }}
+            >
               <DeleteIcon src={icons.deleteIcon} alt="delete-icon" />
             </DeleteButton>
           </ActionsContainer>
@@ -275,6 +308,11 @@ export const UsersTable: FC<TUsersTable> = ({ filtersSelected }) => {
         columns={columns}
         data={users || []}
         noResultsMessage={t('No hay usuarios para mostrar')}
+      />
+      <DeleteConfirmationModal
+        open={isModalOpen}
+        toggleModal={() => setIsModalOpen(false)}
+        idToDelete={idToDelete}
       />
     </TableContainer>
   )