+++ web/usuaris/src/types/types.tsx
@@ -20,14 +20,20 @@ export type TItinerary = {
   slug: string
 }
 
+export type TRole = {
+  id: string
+  name: string
+  slug: string
+}
+
 export type TUpdatedUser = {
   id: string
   dni?: string
   email?: string
   name?: string
   password?: string
-  role?: string
   status?: UserStatus
+  role?: UserRole
   deletedAt?: string
   itineraryId?: string
 }
@@ -36,7 +42,14 @@ export type TUserData = {
   id: string
   name: string
   dni: string
-  status: UserStatus
+  status: string
   createdAt: string
   itineraryName: string
+  role: string
+}
+
+export enum UserRole {
+  ADMIN = 'ADMIN',
+  REGISTERED = 'REGISTERED',
+  MENTOR = 'MENTOR',
 }