+++ web/usuaris/CHANGELOG.md
@@ -2,6 +2,19 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.17.0] - 2024-05-08
+
+### Added
+
+- Added Delete
+
+## [0.16.0] - 2024-05-09
+
+### Added
+
+- Added Role column to UsersTable
+- Added Roles filter feature
+
 ## [0.15.0] - 2024-05-08
 
 ### Added